cleanup: deep-lint round 2 — 443 autofixes + triage annotation for every remaining finding
Fixes: SIM/PT/PLW/RUF autofix sweep (443, incl. 198 stale noqa removals and
56 unneeded int() casts), 8 assert-False -> pytest.fail, 5 explicit
check=False on best-effort subprocess probes (failure tolerated by design),
1 autofix-induced E402 (import hoisted properly).

Triage: every remaining deep finding now carries one of three fates,
documented at the sweep definition in lint.sh —
  FIX-OVER-TIME: PT011/PT017, RUF012, SIM102/108/117, PERF401/403, RUF005/007
  KNOW-AND-BE-CAREFUL: NPY002 (RNG streams), B905 (zip strict semantics),
    B008 (intentional interface defaults)
  WON'T-FIX (ignored with reasons): PLW2901, SIM105, PLW0603 (+ the earlier
    RUF001/2/3, PERF203)
Deep backlog: ~800 raw -> ~190 annotated actionable findings.

Full test suite green; lint.sh --changed clean.

+++ chess_anti_engine/eval/audit.py
@@ -94,11 +94,11 @@ def decode_board_from_planes(
     if root:
         castle_planes = (105, 104, 107, 106)  # us-K, us-Q, them-K, them-Q
         rule50_raw = float(arr[109, 0, 0])
-        rule50 = int(round(rule50_raw * 100.0)) if hist == "lc0_root_legacy_meta" else int(round(rule50_raw))
+        rule50 = round(rule50_raw * 100.0) if hist == "lc0_root_legacy_meta" else round(rule50_raw)
         ep_plane = arr[110] if hist == "lc0_root_legacy_meta" else None
     else:
         castle_planes = (96, 97, 98, 99)
-        rule50 = int(round(float(arr[102, 0, 0]) * 100.0))
+        rule50 = round(float(arr[102, 0, 0]) * 100.0)
         ep_plane = arr[100]
 
     board = chess.Board(fen=None)
